@@ -1,5 +1,7 @@
 module github.com/beeker1121/goque
 
-go 1.13
+go 1.18
 
 require github.com/syndtr/goleveldb v1.0.0
+
+require github.com/golang/snappy v0.0.0-20180518054509-2e65f85255db // indirect
