@@ -30,7 +30,7 @@ func (i *Item) ToString() string {
 // when using this function. This is due to how the encoding/gob
 // package works. Because of this, you should only use this function
 // to decode simple types.
-func (i *Item) ToObject(value interface{}) error {
+func (i *Item) ToObject(value any) error {
 	buffer := bytes.NewBuffer(i.Value)
 	dec := gob.NewDecoder(buffer)
 	return dec.Decode(value)
@@ -42,7 +42,7 @@ func (i *Item) ToObject(value interface{}) error {
 // The value passed to this method should be a pointer to a variable
 // of the type you wish to decode into. The variable pointed to will
 // hold the decoded object.
-func (i *Item) ToObjectFromJSON(value interface{}) error {
+func (i *Item) ToObjectFromJSON(value any) error {
 	return json.Unmarshal(i.Value, value)
 }
 
@@ -70,7 +70,7 @@ func (pi *PriorityItem) ToString() string {
 // when using this function. This is due to how the encoding/gob
 // package works. Because of this, you should only use this function
 // to decode simple types.
-func (pi *PriorityItem) ToObject(value interface{}) error {
+func (pi *PriorityItem) ToObject(value any) error {
 	buffer := bytes.NewBuffer(pi.Value)
 	dec := gob.NewDecoder(buffer)
 	return dec.Decode(value)
@@ -82,7 +82,7 @@ func (pi *PriorityItem) ToObject(value interface{}) error {
 // The value passed to this method should be a pointer to a variable
 // of the type you wish to decode into. The variable pointed to will
 // hold the decoded object.
-func (pi *PriorityItem) ToObjectFromJSON(value interface{}) error {
+func (pi *PriorityItem) ToObjectFromJSON(value any) error {
 	return json.Unmarshal(pi.Value, value)
 }
 
